requirements.txt
# ─── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0

# ─── Database & ORM ────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# ─── Supabase ──────────────────────────────────────────────────────────────────
supabase==2.10.0
postgrest==0.17.1

# ─── Data Validation ───────────────────────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.7.0

# ─── Async HTTP ────────────────────────────────────────────────────────────────
httpx==0.28.1
aiohttp==3.11.11

# ─── Caching ───────────────────────────────────────────────────────────────────
redis==5.2.1
hiredis==3.1.0

# ─── Auth & Security ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# ─── Rate Limiting ─────────────────────────────────────────────────────────────
slowapi==0.1.9

# ─── Logging & Monitoring ──────────────────────────────────────────────────────
structlog==24.4.0
python-json-logger==3.2.1

# ─── Resilience ────────────────────────────────────────────────────────────────
tenacity==9.0.0

# ─── OpenAI ───────────────────────────────────────────────────────────────────
openai==1.82.0

# ─── Audio Processing ─────────────────────────────────────────────────────
pydub==0.25.1

# ─── Cloud Storage (Cloudflare R2) ────────────────────────────────────────────
boto3==1.35.0

# ─── Firebase (Push Notifications) ───────────────────────────────────────────
firebase-admin==6.6.0

# ─── Utilities ─────────────────────────────────────────────────────────────
python-dotenv==1.0.1
orjson==3.10.12
ujson==5.10.0
geopy==2.4.1

# ─── Dev & Testing ─────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
